Auto-PR: extract formatElapsedTime to shared utility (closes #455)
